Refactor rename and move checkbox
