+++ src/components/equipment/CustomCheckBox.jsx
